+++ CHANGES.md
@@ -48,6 +48,7 @@ New:
 - Rename `unsafe.single.infallible` to `single.infallible`.
 - Add `list.indexed`.
 - Added optional support for high-resolution time and latency control on POSIX systems (#1050).
+- Added syntax for `for` and `while` loops (#1252).
 
 Changed: